# docs/openapi.yaml
openapi: 3.1.0
info:
  title: oh-my-commic API
  version: "1.0.0"
  description: >
    面向亲子的 AI 漫画书应用后端 API。所有业务端点位于版本化前缀 /api/v1 下；
    健康探针 GET /api/health 保持不版本化（运维/CI 依赖其稳定路径）。
    认证使用 cookie session（登录后自动 set `session` cookie），前端请求需带
    credentials: 'include'。跨用户/不存在的资源一律返回 404（不泄露存在性）；
    AI/上游错误返回 502（504 超时 / 429 限流）；积分不足返回 402。
    错误响应统一为 `{"error": string}` 信封。图片经 /media/... 直接访问（不在本契约内）。
servers:
  - url: /
...
paths:
...
  /api/v1/books:
    get:
      tags: [books]
      summary: 列出当前用户的书
      responses:
        "200":
          description: 书列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Book"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [books]
      summary: 新建书
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookInput"
      responses:
        "201":
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Book"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/v1/panels/{id}/process:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [story]
      summary: 第 2 段：从 content 解析结构字段 + image_prompt
      responses:
        "200":
          description: 解析后的分镜
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/panels/{id}/render:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [panels]
      summary: 同步生成单格图片（吉卜力风），消耗积分
      description: >
        同步等待生图完成（可能 10-60s），imageUrl 落 /media/{bookId}/..，status=done；
        失败时 status=failed。积分不足返回 402（此时不会调用生图接口）。
      responses:
        "200":
          description: 生图后的分镜
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Panel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "402":
          $ref: "#/components/responses/PaymentRequired"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /api/v1/chapters/{id}/storyboard-chat:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [story]
      summary: 第 1 段：对话拆镜，产出 content-only 分镜并落库
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                messages:
                  type: array
                  items:
                    $ref: "#/components/schemas/ConversationMsg"
                panelCount:
                  type: integer
      responses:
        "200":
          description: 模型回复 + 落库的分镜
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [reply, panels]
                properties:
                  reply:
                    type: string
                  panels:
                    type: array
                    items:
                      $ref: "#/components/schemas/Panel"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
        "502":
          $ref: "#/components/responses/BadGateway"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
  responses:
...
    BadRequest:
      description: 请求参数/格式错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: 未登录
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    PaymentRequired:
      description: 积分不足
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: 资源不存在或跨用户访问（不泄露存在性）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    RateLimited:
      description: 上游限流
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadGateway:
      description: AI/上游服务错误（不泄露上游细节）
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    GatewayTimeout:
      description: 上游超时
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: string
...
    BookInput:
      type: object
      additionalProperties: false
      properties:
        title:
          type: string
        style:
          type: string
        summary:
          type: string
    Book:
      type: object
      additionalProperties: false
      required:
        - id
        - userId
        - title
        - coverUrl
        - style
        - summary
        - isPublic
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
          format: int64
        userId:
          type: integer
          format: int64
        title:
          type: string
        coverUrl:
          type: string
        style:
          type: string
        summary:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ConversationMsg:
      type: object
      additionalProperties: false
      required: [role, content]
      properties:
        role:
          type: string
        content:
          type: string
...
    Panel:
      type: object
      additionalProperties: false
      required:
        - id
        - chapterId
        - order
        - content
        - caption
        - characterIds
        - sceneId
        - imagePrompt
        - imageUrl
        - status
        - location
        - event
        - charExpressions
      properties:
        id:
          type: integer
          format: int64
        chapterId:
          type: integer
          format: int64
        order:
          type: integer
        content:
          type: string
        caption:
          type: string
        characterIds:
          type: [array, "null"]
          items:
            type: integer
            format: int64
        sceneId:
          type: integer
          format: int64
        imagePrompt:
          type: string
        imageUrl:
          type: string
        status:
          type: string
        location:
          type: string
        event:
          type: string
        charExpressions:
          type: [object, "null"]
          additionalProperties:
            type: string
security:
  - sessionCookie: []